% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crpart.R
\name{createpartitions}
\alias{createpartitions}
\title{Create partitions needed for MRA}
\usage{
createpartitions(globaldata, regionaldata, partitions, graphs = F)
}
\arguments{
\item{globaldata}{Data frame object generated by NC2DFG function}

\item{regionaldata}{Data frame object generated by NC2DFR function}

\item{partitions}{A list that specify the way of the algorithm will generate the partitions, the length of the list will be number of levels that the MRA will have.}

\item{graphs}{Logical, show or not the proccess with graphs}
}
\description{
This function creates the object needed
before the Multi Resolution Analisys (MRA)
with global and regional models output
}
\examples{

###Not run
partitions <- list(1,c(1:4),seq(1,2*2*2*11),
  seq(1,2*2*3*2*11),
  seq(1,2*2*3*3*2*11))
length(partitions)

createpartitions(globaldata, regionaldata, partitions, graphs=T)

}
\keyword{2D}
\keyword{Climate}
\keyword{MRA,}
\keyword{data}
\keyword{partitions,}
