% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFun.R
\name{DownloadDataF}
\alias{DownloadDataF}
\title{Download .nc datasets from Narccap}
\usage{
DownloadDataF(table, path = " ", AbName, FromYear = 1968, ToYear = 2071)
}
\arguments{
\item{table}{Is the table produced with the function ShowData()}

\item{path}{The directory where you are going to download the data, we recomend to create a new directory for each variable.}

\item{AbName}{The abbreviate name of the variable you are going to download}

\item{FromYear}{The initial year data sets}

\item{ToYear}{The last year  of data sets}
}
\description{
This function downloads the specified datasets from NARCCAP
}
\examples{
###Not run
table <- ShowData()
DownloadDataF(table=table, path="~/RegionalModelPrecipitation", AbName="pr", FromYear=1968, ToYear=1990)
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
