% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCtoDFfunction.R
\name{NC2DF}
\alias{NC2DF}
\title{DESCRIPTION OF THE FUNCTION}
\usage{
NC2DF(PATH)
}
\arguments{
\item{PATH}{Is the directory where is located the .nc files of a same variable}
}
\description{
DESCRIPTION OF THE FUNCTION
}
\examples{
ShowData()

}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
