% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCtoDFGlobalFunction.R
\name{NC2DFG}
\alias{NC2DFG}
\title{This  function coverts the .nc downloaded  from NARCCAP  in a data.frame object.}
\usage{
NC2DFG(PATH, VARNAME)
}
\arguments{
\item{PATH}{Is the directory where is located the .nc files of a same variable}

\item{VARNAME}{String, is the name of the variable in the dataset}
}
\description{
This proccess may require high RAM capacity
}
\examples{
NC2DF("~/PrecipitationGlobalModelData")
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
