% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{VisualizeDATA}
\alias{VisualizeDATA}
\title{VisualizeDATA function generates a map or a plot of the chosen variable in a specific month and year}
\usage{
VisualizeDATA(datos, var, year, month, type)
}
\arguments{
\item{var}{Name of the variable chosen to be visualized}

\item{year}{A year from 1968 to 1999}

\item{month}{Number of the month you want}

\item{data}{Data frame objet generated with NC2DF function}
}
\description{
VisualizeDATA function generates a map or a plot of the chosen variable in a specific month and year
}
\keyword{Climate}
\keyword{Map,}
\keyword{Visualization,}
\keyword{data}
