% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFun.R
\name{DownloadDataF}
\alias{DownloadDataF}
\title{Show Narccap datasets available to download}
\usage{
DownloadDataF(table, path = " ", AbName, FromYear = 1968, ToYear = 2071)
}
\description{
This function downloads the specified datasets
}
\examples{
ShowData()
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
