% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoiningFunction.R
\name{joinRGdata}
\alias{joinRGdata}
\title{Join the regional and global datasets}
\usage{
joinRGdata(var, varreg)
}
\arguments{
\item{var}{is the dataframe produced by the function NC2DFG()}

\item{varreg}{is the dataframe produced by the function NC2DFG()}
}
\description{
This function joins the processed data from regional and global model in a Data Frame that matches the points using grids
}
\examples{

###Not run
joinRGdata(globaldata, regionaldata)
}
\keyword{data}
\keyword{global}
\keyword{joining,}
\keyword{model,}
\keyword{regional}
