% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoiningFunction.R
\name{joinRGdata}
\alias{joinRGdata}
\title{DESCRIPTION OF THE FUNCTION}
\usage{
joinRGdata(var, varreg)
}
\arguments{
\item{var}{is the dataframe produced by the function NC2DFG()}

\item{varreg}{is the dataframe produced by the function NC2DFG()}
}
\description{
DESCRIPTION OF THE FUNCTION
}
\examples{
NC2DFR("~/RegionalModelPrecipitation")
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
