% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationCount.R
\name{narccapCOUNT}
\alias{narccapCOUNT}
\title{Plot a comparison count of data observations}
\usage{
narccapCOUNT(data, year, month)
}
\arguments{
\item{data}{Data frame object generated by joinRGdata function}

\item{year}{A year of the dataset}

\item{month}{Number of the month}
}
\description{
narccapCount function generates a graph
that compares the number of global points
with the number of regional points
}
\examples{

###Not run
joinRGdata(globaldata, regionaldata)
narccapCOUNT(JoinnedDF, 1976, 11)

}
\keyword{Climate}
\keyword{Visualization,}
\keyword{data}
