% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{narccapMAP}
\alias{narccapMAP}
\title{narccapMAP function generates a map of the chosen variable in a specific month and year}
\usage{
narccapMAP(datos, var, year, month)
}
\arguments{
\item{var}{Name of the variable chosen to be visualized}

\item{year}{A year from 1968 to 1999}

\item{month}{Number of the month you want}

\item{data}{Data frame objet generated with NC2DF function}
}
\description{
narccapMAP function generates a map of the chosen variable in a specific month and year
}
\keyword{Climate}
\keyword{Map,}
\keyword{Visualization,}
\keyword{data}
