% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCtoDFGlobalFunction.R
\name{NC2DF}
\alias{NC2DF}
\title{This  function coverts the .nc downloaded  from NARCCAP  in a data.frame object.
This proccess may require high RAM capacity}
\usage{
NC2DF(PATH, TYPE)
}
\arguments{
\item{PATH}{Is the directory where is located the .nc files of a same variable}

\item{TYPE}{Is the origin of the data, set TYPE="R" if origin  is the regional  model or TYPE="G" if the  origin is  the global model}
}
\description{
This  function coverts the .nc downloaded  from NARCCAP  in a data.frame object.
This proccess may require high RAM capacity
}
\examples{
NC2DF("~/PrecipitationGlobalModelData",TYPE="G")
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
