% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCtoDFGlobalFunction.R
\name{NC2DFG}
\alias{NC2DFG}
\title{DESCRIPTION OF THE FUNCTION}
\usage{
NC2DFG(PATH)
}
\arguments{
\item{PATH}{Is the directory where is located the .nc files of a same variable}
}
\description{
DESCRIPTION OF THE FUNCTION
}
\examples{
NC2DFG("~/PrecipitationGlobalModelData")

}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
