% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCtoDFfunction.R
\name{NC2DFR}
\alias{NC2DFR}
\title{Convert the narccap nc file in a data frame}
\usage{
NC2DFR(PATH)
}
\arguments{
\item{PATH}{Is the directory where is located the .nc files of a same variable}
}
\description{
This  function coverts the .nc downloaded  from NARCCAP regional models in a data.frame object.
This proccess may require high RAM memory capacity
}
\examples{
NC2DFR("~/RegionalModelPrecipitation")
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
