% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCtoDFfunction.R
\name{NC2DFR}
\alias{NC2DFR}
\title{DESCRIPTION OF THE FUNCTION
asdasfasgas}
\usage{
NC2DFR(PATH)
}
\arguments{
\item{PATH}{Is the directory where is located the .nc files of a same variable}
}
\description{
DESCRIPTION OF THE FUNCTION
asdasfasgas
}
\examples{
NC2DFR("~/RegionalModelPrecipitation")
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
