% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showdatafunction.R
\name{ShowData}
\alias{ShowData}
\title{Show Narccap datasets available to download}
\usage{
ShowData()
}
\description{
This function creates a table with the information of the diferent datasets
available to download from R in NARCCAP web site.
}
\examples{
ShowData()
}
\keyword{data}
\keyword{datasets,}
\keyword{download,}
